actions for pay created, card in preview
