--- client/src/components/Paypal/Preview/CardPreview.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const CardPreview = () => {
-  return (
-    <>
-      <h1>CardPreview</h1>
-    </>
-  );
-};
-
-export default CardPreview;
